Actually got GameNetworkingsockets to work! Progress on re-writing the Client/Server interfaces
