+++ .gitignore
@@ -2,3 +2,4 @@
 .ipch
 out
 build
+vcpkg_installed
