+++ Pyxis/src/Pyxis/Network/Network.h
@@ -1,9 +1,7 @@
 #pragma once
 
-#include "NetworkCommon.h"
 #include "NetworkMessage.h"
 #include "NetworkThreadSafeQueue.h"
-#include "NetworkConnection.h"
 #include "NetworkClient.h"
 #include "NetworkServer.h"
