+++ Pyxis/src/Pyxis/Network/NetworkClient.h
@@ -1,9 +1,7 @@
 #pragma once
 
-#include "NetworkCommon.h"
 #include "NetworkMessage.h"
 #include "NetworkThreadSafeQueue.h"
-#include "NetworkConnection.h"
 
 namespace Pyxis
 {
