--- Pyxis/src/Pyxis/Network/NetworkCommon.cpp
@@ -1,3 +1,0 @@
-#include "pxpch.h"
-
-#include "NetworkCommon.h"
