--- Pyxis/src/Pyxis/Network/NetworkConnection.cpp
@@ -1,3 +1,0 @@
-#include "pxpch.h"
-
-#include "NetworkConnection.h"
