--- Pyxis/src/Pyxis/Network/SteamNetworkClient.h
@@ -1,21 +1,0 @@
-#pragma once
-
-#include <steam/isteamnetworkingsockets.h>
-#include <steam/isteamnetworkingutils.h>
-
-
-namespace Pyxis
-{
-	class SteamServer
-	{
-	public:
-
-		
-
-	private:
-
-		HSteamNetConnection m_hConnection;
-		ISteamNetworkingSockets* m_pInterface;
-
-	};
-}
