+++ vcpkg.json
@@ -1,0 +1,9 @@
+{
+  "dependencies": [
+    {
+      "name": "gamenetworkingsockets",
+      "version>=": "1.4.1"
+    }
+  ],
+  "builtin-baseline": "9b5cb8e554487f3edb5d50b80188883846f81e14"
+}
